# iron-spot-api/src/main/java/com/ironspot/gym/GymController.java
package com.ironspot.gym;

import com.ironspot.common.dto.ErrorResponse;
import com.ironspot.common.exception.BusinessException;
import com.ironspot.gym.dto.GymDetailResponse;
import com.ironspot.gym.dto.GymSearchRequest;
import com.ironspot.gym.dto.GymWithMachineCountResponse;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import org.springdoc.core.annotations.ParameterObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.ModelAttribute;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;
import java.util.UUID;

@RestController
@RequestMapping(value = "/api/gyms", produces = MediaType.APPLICATION_JSON_VALUE)
@RequiredArgsConstructor
public class GymController {

    private final GymService gymService;

    @GetMapping("/search")
    @Operation(summary = "Search gyms within map bounds", tags = {"gyms"})
    @ApiResponses({
        @io.swagger.v3.oas.annotations.responses.ApiResponse(
            responseCode = "200", description = "Gym list returned successfully"),
        @io.swagger.v3.oas.annotations.responses.ApiResponse(
            responseCode = "400", description = "Missing or invalid bounds parameters",
            content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
    })
    public List<GymWithMachineCountResponse> search(
        @Valid @ModelAttribute @ParameterObject GymSearchRequest request
    ) {
        return gymService.searchInBounds(request);
    }

    @GetMapping("/{id}")
    @Operation(summary = "Get gym detail", tags = {"gyms"})
    @ApiResponses({
        @io.swagger.v3.oas.annotations.responses.ApiResponse(
            responseCode = "200", description = "Gym returned successfully"),
        @io.swagger.v3.oas.annotations.responses.ApiResponse(
            responseCode = "404", description = "Gym not found",
            content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
    })
    public GymDetailResponse getById(@PathVariable UUID id) {
        return gymService.findById(id)
            .orElseThrow(() -> new BusinessException("Gym not found: " + id, HttpStatus.NOT_FOUND));
    }
}