# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IronSpot API",
    "description": "헬스장 기구 정보 플랫폼 API",
    "version": "v1"
  },
  "servers": [{ "url": "http://localhost:8080", "description": "Generated server url" }],
  "security": [{ "Bearer": [] }],
  "paths": {
    "/api/machines/{gymMachineId}/photos": {
      "get": {
        "tags": ["photos"],
        "summary": "List photos for a gym machine",
        "operationId": "listPhotos",
        "parameters": [
          {
            "name": "gymMachineId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "200": {
            "description": "Photo list returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/PhotoResponse" }
                }
              }
            }
          }
        }
      }
    },
    "/api/gyms/{id}": {
      "get": {
        "tags": ["gyms"],
        "summary": "Get gym detail",
        "operationId": "getById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "200": {
            "description": "Gym returned successfully",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/GymDetailResponse" } }
            }
          },
          "404": {
            "description": "Gym not found",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          }
        }
      }
    },
    "/api/gyms/{gymId}/machines": {
      "get": {
        "tags": ["machines"],
        "summary": "List machines in a gym",
        "operationId": "listMachines",
        "parameters": [
          {
            "name": "gymId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "200": {
            "description": "Machine list returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/GymMachineResponse" }
                }
              }
            }
          }
        }
      }
    },
    "/api/gyms/search": {
      "get": {
        "tags": ["gyms"],
        "summary": "Search gyms within map bounds",
        "operationId": "search",
        "parameters": [
          {
            "name": "minLat",
            "in": "query",
            "required": true,
            "schema": { "type": "number", "format": "double", "maximum": 90.0, "minimum": -90.0 }
          },
          {
            "name": "maxLat",
            "in": "query",
            "required": true,
            "schema": { "type": "number", "format": "double", "maximum": 90.0, "minimum": -90.0 }
          },
          {
            "name": "minLng",
            "in": "query",
            "required": true,
            "schema": { "type": "number", "format": "double", "maximum": 180.0, "minimum": -180.0 }
          },
          {
            "name": "maxLng",
            "in": "query",
            "required": true,
            "schema": { "type": "number", "format": "double", "maximum": 180.0, "minimum": -180.0 }
          },
          { "name": "brandId", "in": "query", "required": false, "schema": { "type": "string" } },
          {
            "name": "categoryId",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "loadingType",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "200": {
            "description": "Gym list returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/GymWithMachineCountResponse" }
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid bounds parameters",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          }
        }
      }
    },
    "/api/categories": {
      "get": {
        "tags": ["categories"],
        "summary": "List all categories",
        "operationId": "listCategories",
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "200": {
            "description": "Category list returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/CategoryResponse" }
                }
              }
            }
          }
        }
      }
    },
    "/api/brands": {
      "get": {
        "tags": ["brands"],
        "summary": "List all brands",
        "operationId": "listBrands",
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "200": {
            "description": "Brand list returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/BrandResponse" }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"]
      },
      "PhotoResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "gymMachineId": { "type": "string", "format": "uuid" },
          "userId": { "type": "string", "format": "uuid" },
          "photoUrl": { "type": "string" },
          "upvoteCount": { "type": "integer", "format": "int32" },
          "createdAt": { "type": "string", "format": "date-time" }
        },
        "required": ["createdAt", "gymMachineId", "id", "photoUrl", "upvoteCount", "userId"]
      },
      "GymDetailResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "address": { "type": "string" },
          "latitude": { "type": "number", "format": "double" },
          "longitude": { "type": "number", "format": "double" },
          "phone": { "type": "string" },
          "operatingHours": { "type": "string" },
          "dayPassPrice": { "type": "integer", "format": "int32" },
          "isVerified": { "type": "boolean" },
          "lastVerifiedAt": { "type": "string", "format": "date-time" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "required": [
          "address",
          "createdAt",
          "id",
          "isVerified",
          "latitude",
          "longitude",
          "name",
          "updatedAt"
        ]
      },
      "GymMachineResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "quantity": { "type": "integer", "format": "int32" },
          "isCustom": { "type": "boolean" },
          "customName": { "type": "string" },
          "lastVerifiedAt": { "type": "string", "format": "date-time" },
          "templateId": { "type": "string", "format": "uuid" },
          "machineName": { "type": "string" },
          "loadingType": { "type": "string" },
          "brandId": { "type": "string", "format": "uuid" },
          "brandName": { "type": "string" },
          "categoryId": { "type": "string", "format": "uuid" },
          "categoryName": { "type": "string" },
          "photoCount": { "type": "integer", "format": "int64" }
        },
        "required": ["id", "isCustom", "photoCount", "quantity"]
      },
      "GymWithMachineCountResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "address": { "type": "string" },
          "latitude": { "type": "number", "format": "double" },
          "longitude": { "type": "number", "format": "double" },
          "phone": { "type": "string" },
          "operatingHours": { "type": "string" },
          "dayPassPrice": { "type": "integer", "format": "int32" },
          "isVerified": { "type": "boolean" },
          "lastVerifiedAt": { "type": "string", "format": "date-time" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "machineCount": { "type": "integer", "format": "int64" }
        },
        "required": [
          "address",
          "createdAt",
          "id",
          "isVerified",
          "latitude",
          "longitude",
          "machineCount",
          "name",
          "updatedAt"
        ]
      },
      "CategoryResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" }
        },
        "required": ["id", "name"]
      },
      "BrandResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" }
        },
        "required": ["id", "name"]
      }
    },
    "securitySchemes": { "Bearer": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" } }
  }
}