# src/shared/generated/model/errorResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * IronSpot API
 * 헬스장 기구 정보 플랫폼 API
 * OpenAPI spec version: v1
 */

export interface ErrorResponse {
  error: string;
}