# src/shared/generated/model/gymDetailResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * IronSpot API
 * 헬스장 기구 정보 플랫폼 API
 * OpenAPI spec version: v1
 */

export interface GymDetailResponse {
  id: string;
  name: string;
  address: string;
  latitude: number;
  longitude: number;
  phone?: string;
  operatingHours?: string;
  dayPassPrice?: number;
  isVerified: boolean;
  lastVerifiedAt?: string;
  createdAt: string;
  updatedAt: string;
}