# src/shared/generated/model/gymWithMachineCountResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * IronSpot API
 * 헬스장 기구 정보 플랫폼 API
 * OpenAPI spec version: v1
 */

export interface GymWithMachineCountResponse {
  id: string;
  name: string;
  address: string;
  latitude: number;
  longitude: number;
  phone?: string;
  operatingHours?: string;
  dayPassPrice?: number;
  isVerified: boolean;
  lastVerifiedAt?: string;
  createdAt: string;
  updatedAt: string;
  machineCount: number;
}